+++ src/Migrations/Version20251210115500.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Migrations;
 
-use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Entity\CachedCourse;
-use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Entity\CachedCourseTitle;
 use Doctrine\DBAL\Schema\Schema;
 
 final class Version20251210115500 extends EntityManagerMigration
@@ -17,14 +15,6 @@ final class Version20251210115500 extends EntityManagerMigration
 
     public function up(Schema $schema): void
     {
-        $coursesTableName = CachedCourse::TABLE_NAME;
-        $coursesStagingTableName = CachedCourse::STAGING_TABLE_NAME;
-
-        $this->addSql("CREATE TABLE $coursesStagingTableName LIKE $coursesTableName");
-
-        $courseTitlesTableName = CachedCourseTitle::TABLE_NAME;
-        $courseTitlesStagingTableName = CachedCourseTitle::STAGING_TABLE_NAME;
-
-        $this->addSql("CREATE TABLE $courseTitlesStagingTableName LIKE $courseTitlesTableName");
+        $this->recreateCacheTables();
     }
 }